requirements.txt
mysql-connector-python==8.0.33
streamlit==1.22.0
pandas==1.5.3 